src/main/java/com/github/brunomndantas/jscrapper/support/parser/collection/text/reference/CollectionByteTextParser.java
package com.github.brunomndantas.jscrapper.support.parser.collection.text.reference;

import com.github.brunomndantas.jscrapper.support.parser.collection.text.CollectionTextParser;
import org.openqa.selenium.WebElement;

public class CollectionByteTextParser extends CollectionTextParser {

    public CollectionByteTextParser() {
        super(null);
    }



    @Override
    protected Object parseElement(WebElement element) throws Exception {
        return Byte.parseByte(element.getText());
    }

}